add Players.js to better organize code
