+++ index.html
@@ -8,7 +8,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=MedievalSharp&display=swap" rel="stylesheet">
-    <script src="./script.js" defer></script>
+    <script src="./script.js" type="module" defer></script>
     <title>World of Codecraft</title>
 </head>
 <body>
